requirements.txt

biopython
numpy
pandas
matplotlib
seaborn
